Feat: notification data loading operations
The functionality of opening cont notifications with type "=handleShowContNotifications"
has been completed, in which the following actions were performed: when opening the notification icon,
if there are unseen notifications, then first of all, the ID is written to the object according
to the current cycle, then the array is overwritten from action to an array with a hook, the same actions
are performed to open the existing viewed ones notifications. Also, when opened, the notification counter is reset,
which is used from the context hook, which counts the available number of unread notifications. Further, after closing,
both sections are overwritten from unread notifications to viewed ones, in addition, the changed state of the object
field "isViewedNotifcation: true" is added, new data of viewed notifications is also saved to action and to the db collection,
in addition, action and data from the database about unread notifications are cleared using the corresponding dispatches.