Feat: checking for notification receipt
Adding notification receipt processing after checking the inform error, if the user has added
the notification type to the blocked ones, then the notification will process the verified inform error,
process its data, but will not display it. Accordingly, if the user did not add to the blocked ones,
then the corresponding operations will take place and the notification will be displayed.