Feat: operation data processing notification events
Processing the data of the event to restore the types of blocked notifications,
after the user has clicked on the action button to hide or block the notification.
Further, after processing these actions, the user wishes to return the data of the blocked
notifications, unblocking of these actions is available, which are given in the corresponding block of the functionality code.