Feat: operation displaying data notifications
A functional has been made in which notification data is displayed,
such as: the notification text, the date and its formatting
of the value passed to action, the display of the icon of the message recipient.
The array with the notification goes by the minus index, i.e. from the end of the array,
this was done so that new notifications would be displayed first. Also, functionality
has been added and displayed by clicking on the arrow when hovering, blocking the notification,
then prohibiting receipt of notifications from one or another recipient; hiding the notification.
It is possible to restore the existing blocked notifications, their recovery method and functional
description were presented in an early commit - "259fcc0". In addition, in this functionality,
operations are performed with the addition of blocked or hidden notifications, adding to the
user's database using the appropriate dispatches in "userSlice.js", commit is " 91a9586".