Feat: added slice - "filteredHistory"
The slice was created to work with these notification display types
and filter them, as well as to conveniently view selected notifications
on the notification history page. In addition, the following redirectors were created:
"setActionFilteredNots" - an array for storing notification type data;
"setFilterActionNots" - array filtering, the user can hide by clicking,
choosing the desired type to view the notification history with the selected parameter;
"setRemoveFilterActs" - to clear the array data, used when logging out of the account,
because the array "arrActionFilteredNots" can only be processed if the user
is registered or logged in to the account, as well as the browser is used for viewing,
an option when the user needs to view the history of all notifications without filtering.
Corresponding exports of these reducers have also been added for use in other functional components.