Feat: added operations filter notifications
Functionality for adding a filter that contains the following types of notifications:
SS Coorchik - Support Service, Coorchik - Admin, Couriers. This filtering was done
so that you can sort the notifications you want to view. To save the filtering data,
the following processor was used - "setActionFilteredNots", which was widely described in the commit - "a83ed1a".