Feat: slice - "notificationsAgreement"
A slice for storing and managing user data about notification settings.
The user can disable and disable notifications with the following variables,
which are used in the reduxer - "setUserNotificationsSettings":
the variable "notifications" is a global parameter in the notification settings;
"multipleNotifications" is a multi-notification parameter, that is,
the notification occurs in real time; "notificationsSs" is a notification parameter
from Support Service Coorchik; "notificationsAdmin" is a notification parameter from Coorchik;
"notificationsCouriers" is a notification parameter from Couriers.
In addition, reduxers have been created to update these variables.
Also, a redeser has been created to clear notification settings when you log out of your account.