Feat: use update slice data - "notificationsAgreementSlice"
The dispatches for updating these parameters with the prefix:
"setUpd" were performed for two actions in the program:
first, after the user's actions of the notification type ban - "notificationsSs",
assigning the value - false; secondly, after the user reinstates the banned notifications,
in this case, the unbanking of all banned notifications by the reduxers is used:
"setUpdNotificationsSsSettings" , "setUpdNotificationsAdminSettings", "setUpdNotificationsCouriersSettings".
The purpose of the sliceis "notificationsAgreement" and its editors, described in the commit - "4c839e6".