Feat: operations after hide all filters
Added functionality in which, if the user has hidden a filter that has a notification,
then in this case a hint is displayed saying that the user can add a filter
that has notifications. When you click on the text, it can be:
"SS Coorchik", "Coorchik", "Couriers". When pressed, the cont opens to add a filter,
and the filter that has notifications will play an animation in order to make it clear to the user to choose the type.