Feat: added operations and displaying - "manageNotifications"
The following actions were performed with the block component
on the "notificationsHistory and Manage" page: the layout
of the block was added, which has the names of each type of setting,
as well as its explanation;operations were added to configure the notification by the user,
that is, permission/refusal to receivegeneral notifications or from specific addressees,
in addition, using data from the created reducer - "notificationsAgreement" (see commit - 4c839e6).
Also, when the notification settings are changed, data is dispensed and updated
in the collection of the user's document in the database - "usersNotifications".