Feat: variable operations data slice - "notificationsAgreement"
A variable has been added that contains a value about the selected settings
of hidden notifications by the user, which implies data
where the user can configure the setting so that hidden notifications:
firstly, they are displayed in the blocks of the notification site,
where it would be mentioned that the user has hidden notifications;
secondly, the number of hidden notifications; thirdly, if the user has hidden notifications,
then he will receive notifications that they exist and the user can perform
subsequent actions with them, this is their viewing or clearing.
Also, a redesigner was added - "setUpdHideNotificationsSettings", to change and manage the user's settings.