Feat: added field - "hiddenNotifications" in create doc
The "hiddenNotifications" field has been added
to the history notifications creation document
after the user has registered. Also, the corresponding
dispatcher of the redeser is executed - "setUserNotificationsSettings".
The default value of "hiddenNotifications" at the time of creating
a user account is "Always", this use of the variable is described in commit - 9a4705d.